get stair indexes to stair model
