+++ analytics/analytics/models/drop_model.py
@@ -32,9 +32,8 @@ class StairModelState(ModelState):
 class DropModel(StairModel):
 
     def get_model_type(self) -> (str, bool):
-        model = 'drop'
         type_model = False
-        return (model, type_model)
+        return (ModelName.DROP, type_model)
 
     def find_segment_center(self, dataframe: pd.DataFrame, start: int, end: int) -> int:
         data = dataframe['value']
@@ -42,17 +41,17 @@ class DropModel(StairModel):
         segment_center_index = utils.find_pattern_center(segment, start, 'drop')
         return segment_center_index
 
-    def get_stair_indexes(self, data: pd.Series, height: float, length: int) -> List[int]:
-        '''
-        data: data, that contains drop segments,
-        length: the number of indexes to be contained in the drop segment,
-        height: the difference between drop max_line and min_line(see utils.find_parameters),
-        return: list of start drop segment indexes
-        '''
-        #TODO: refactor and move method to stair_model
-        indexes = []
-        for i in range(len(data) - length - 1):
-            for x in range(1, length):
-                if(data[i + x] < data[i] - height):
-                    indexes.append(i)
-        return indexes
+    # def get_stair_indexes(self, data: pd.Series, height: float, length: int) -> List[int]:
+    #     '''
+    #     data: data, that contains drop segments,
+    #     length: the number of indexes to be contained in the drop segment,
+    #     height: the difference between drop max_line and min_line(see utils.find_parameters),
+    #     return: list of start drop segment indexes
+    #     '''
+    #     #TODO: refactor and move method to stair_model
+    #     indexes = []
+    #     for i in range(len(data) - length - 1):
+    #         for x in range(1, length):
+    #             if(data[i + x] < data[i] - height):
+    #                 indexes.append(i)
+    #     return indexes
